+++ tests/Common/Models/Transfer.php
@@ -2,6 +2,6 @@
 
 namespace Evinkuraga\Wallet\Test\Common\Models;
 
-class Transfer extends \Bavix\Wallet\Models\Transfer
+class Transfer extends \Evinkuraga\Wallet\Models\Transfer
 {
 }