+++ tests/Common/Models/Wallet.php
@@ -2,6 +2,6 @@
 
 namespace Evinkuraga\Wallet\Test\Common\Models;
 
-class Wallet extends \Bavix\Wallet\Models\Wallet
+class Wallet extends \Evinkuraga\Wallet\Models\Wallet
 {
 }